requirements.txt
# AWS Bedrock (using direct SDK instead of Strands)
boto3>=1.35.0
botocore>=1.35.0

# Strands Agents SDK with bidirectional streaming support (Nova Sonic)
strands-agents[bidi]==1.24.0
strands-agents-evals

# Audio processing
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.11.3
pyyaml>=6.0.0

# Web server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Security constraints (transitive dep minimum versions)
cryptography>=46.0.5
pillow>=12.1.1

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
